Improve RedMemory main buffer check

+++ src/RedSound/RedMemory.cpp
@@ -59,7 +59,7 @@ int RedNew(int size)
 
 	if (size >= 1) {
 		if (m_MemoryBank != 0) {
-			if (m_DataBuffer != 0) {
+			if ((unsigned int)m_DataBuffer != 0) {
 				interrupts = OSDisableInterrupts();
 				address = m_DataBuffer;
 				size = (size + 0x1F) & 0xFFFFFFE0;
